requirements.txt
absl-py
addict==2.4.0
aiofiles==0.8.0
albumentations==1.1.0
asynctest==0.13.0
attrs==21.4.0
braceexpand==0.1.7
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi
charset-normalizer==2.0.10
click==7.1.2
cmake==3.22.4
codecov==2.1.12
colorama==0.4.4
coverage==6.3.2
cryptography
cycler==0.11.0
Cython==0.29.28
dlib==19.23.1
docker-pycreds==0.4.0
faiss==1.7.2
filelock==3.6.0
flake8==4.0.1
fonttools==4.28.5
fvcore==0.1.5.post20220414
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.4.0
idna
imageio==2.13.5
imgaug==0.4.0
importlib-metadata==4.10.0
iniconfig==1.1.1
instaboostfast==0.1.2
iopath==0.1.9
isort==5.10.1
jarowinkler==1.0.2
joblib==1.1.0
kiwisolver==1.3.2
kwarray==0.6.0
lanms-neo==1.0.2
llvmlite==0.38.0
lmdb==1.3.0
Markdown
matplotlib==3.5.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmpycocotools==12.0.3
model-index==0.1.11
networkx==2.6.3
numba==0.55.1
numpy
oauthlib==3.1.1
olefile
opencv-python==4.5.5.62
opencv-python-headless==4.5.4.60
openmim==0.1.5
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==9.1.0
pluggy==1.0.0
portalocker==2.4.0
prettytable==3.2.0
promise==2.3
protobuf==3.19.1
psutil==5.9.0
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser
pyflakes==2.4.0
pyOpenSSL
pyparsing==3.0.6
PySocks
pytest==7.1.1
pytest-cov==3.0.0
pytest-runner==6.0.0
python-dateutil==2.8.2
pytorch-metric-learning==1.3.0
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
rapidfuzz==2.0.9
regex==2022.3.15
requests
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.49
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.5.8
setproctitle==1.2.2
Shapely==1.8.0
shortuuid==1.0.8
simplejson==3.17.6
six
smmap==5.0.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
termcolor==1.1.0
terminaltables==3.1.10
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.11.6
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.3.1
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.62.3
transformers==4.17.0
typing_extensions
ubelt==1.0.1
urllib3
wandb==0.12.11
wcwidth==0.2.5
webdataset==0.2.4
Werkzeug
xdoctest==1.0.0
yacs==0.1.8
yapf==0.32.0
yaspin==2.1.0
zipp
